use the timestamp formatting everywhere and fix reminders

+++ commands/remind.js
@@ -1,6 +1,7 @@
 /**
  * Created by Peter on 01/07/2017.
  */
+const Discord = require('discord.js');
 module.exports = {
     name: "Reminders",
     usage: "remind :command? :args?+",
@@ -72,33 +73,49 @@ module.exports = {
 
         }
     },
-    sendReminder: async function(reminder, bot){
-        if(bot.drain)return;
-        if(!module.exports.deletedReminders.includes(reminder.id)) {
+    buildEmbed(reminder){
+        const embed = new Discord.MessageEmbed();
+        embed.setColor("#6868f5")
+        embed.setTitle(":stopwatch: Reminder");
+        embed.setDescription(`<t:${Math.floor(reminder.timestamp/1000)}:R> you told me to remind you of this:\n>>> \n${reminder.message}\n\u200B`);
+        embed.setTimestamp(reminder.at);
+        return embed;
+    },
+    sendReminder: async function (reminder, bot) {
+        if (bot.drain) return;
+        if (!module.exports.deletedReminders.includes(reminder.id)) {
             bot.logger.log(`Reminding ${reminder.id}: ${reminder.user}: ${reminder.message}`);
             try {
-                if(!reminder.server)throw new Error("DM reminder");
+                if (!reminder.server) throw new Error("DM reminder");
                 const channel = await bot.client.channels.fetch(reminder.channel);
-                await channel.send(await bot.lang.getTranslation(channel.guild.id, "REMIND_REMINDER", {
-                    username: reminder.user,
-                    date: new Date(reminder.timestamp).toString(),
-                    message: reminder.message
-                }));
-            } catch (e) {
-                if(reminder.server) {
-                    bot.logger.log("Reminder channel no longer exists, attempting to send it to the user instead...");
-                }else{
-                    bot.logger.log("DM Reminder, finding user...");
-                }
+                let embed = module.exports.buildEmbed(reminder);
                 try {
+                    await channel.send({
+                        embeds: [module.exports.buildEmbed(reminder)],
+                        reply: {messageReference: reminder.messageID},
+                    });
+                }catch(e){
+                    await channel.send({
+                        content: `<@${reminder.user}>`,
+                        embeds: [module.exports.buildEmbed(reminder)],
+                    });
+                }
+                if(!channel.members.has(reminder.user)){
+                    embed.addField(":warning: Note", "You are receiving this reminder because you are no longer in the channel in which the reminder was set.");
+                    const targetUser = await bot.client.users.fetch(reminder.user);
+                    const dm = await targetUser.createDM();
+                    await dm.send({embeds: [embed]});
+                }
+            } catch (e) {
+                try{
+                    bot.logger.log("Reminder channel no longer exists or is DM, attempting to send it to the user instead...");
                     const targetUser = await bot.client.users.fetch(reminder.user);
                     if (targetUser) {
                         const dm = await targetUser.createDM();
-                        await dm.send(await bot.lang.getTranslation(reminder.channel, "REMIND_REMINDER", {
-                            username: reminder.user,
-                            date: new Date(reminder.timestamp).toString(),
-                            message: reminder.message
-                        }));
+                        let embed = module.exports.buildEmbed(reminder);
+                        if(reminder.server)
+                            embed.addField(":warning: Note", `This reminder was supposed to go to <#${reminder.channel}>, but I couldn't access it. It could be deleted, or I could have been kicked from that server.`);
+                        await dm.send({content: `<@${reminder.user}>`, embeds: [module.exports.buildEmbed(reminder)]});
                     } else {
                         bot.logger.log("Couldn't retrieve the user.");
                     }
@@ -107,13 +124,14 @@ module.exports = {
                     bot.raven.captureException(e);
                 }
             }
-        }else{
+        } else {
             bot.logger.log(`Reminder ${reminder.id} was deleted.`);
         }
-        try{
+        try {
+            if(!reminder.id)return bot.logger.warn("Reminder had no ID");
             await bot.database.removeReminder(reminder.id);
             bot.logger.log(`Removed reminder ${reminder.id}`);
-        }catch(err){
+        } catch (err) {
             bot.logger.error(`Error removing reminder!!! This is bad!!! ${err.stack}`);
             bot.raven.captureException(err);
         }
